7/3 오후작업 ( 입찰 아직 미구현)
bid DAO등 등 의 파일등 수정해야함

+++ src/main/java/com/auction/dao/MemberDAO.java
@@ -43,7 +43,7 @@ public class MemberDAO {
 	            loginUser.setBirthdate(rs.getString("BIRTHDATE"));
 	            loginUser.setGender(rs.getString("GENDER"));
 	            loginUser.setMobileCarrier(rs.getString("MOBILE_CARRIER"));
-	            loginUser.setMileage(rs.getInt("MILEAGE"));
+	            loginUser.setMileage(rs.getLong("MILEAGE"));
 	            loginUser.setZip(rs.getString("ZIP"));
 	            loginUser.setAddr1(rs.getString("ADDR1"));
 	            loginUser.setAddr2(rs.getString("ADDR2"));
